chore: update tabs to remove console logs
