+++ src/renderer/src/routes/(MapTabs)/_Map.tsx
@@ -1,28 +1,10 @@
 import { Suspense } from 'react'
-import InfoOutlinedIcon from '@mui/icons-material/InfoOutlined'
-import PostAddIcon from '@mui/icons-material/PostAdd'
-import SettingsIcon from '@mui/icons-material/Settings'
-import { Box, CircularProgress, Paper } from '@mui/material'
-import Tab from '@mui/material/Tab'
-import Tabs from '@mui/material/Tabs'
+import { CircularProgress, Paper } from '@mui/material'
 import { styled } from '@mui/material/styles'
-import { Outlet, createFileRoute, useNavigate } from '@tanstack/react-router'
-import { defineMessages, useIntl } from 'react-intl'
+import { Outlet, createFileRoute } from '@tanstack/react-router'
 
 import { VERY_LIGHT_GREY, WHITE } from '../../colors'
-import { Text } from '../../components/Text'
-import type { FileRoutesById } from '../../routeTree.gen'
-
-const m = defineMessages({
-	setting: {
-		id: 'tabBar.label.settings',
-		defaultMessage: 'Settings',
-	},
-	about: {
-		id: 'tabBar.label.about',
-		defaultMessage: 'About',
-	},
-})
+import { Tabs } from '../../components/Tabs'
 
 const Container = styled('div')({
 	display: 'flex',
@@ -30,55 +12,17 @@ const Container = styled('div')({
 	height: '100%',
 })
 
-const MapTabStyled = styled(MapTab)({
-	width: 60,
-})
-
 export const Route = createFileRoute('/(MapTabs)/_Map')({
 	component: MapLayout,
 })
 
 export function MapLayout() {
-	const navigate = useNavigate()
-	const { formatMessage } = useIntl()
 	return (
 		<Container>
 			<Paper elevation={3} sx={{ display: 'flex' }}>
-				<Tabs
-					sx={{ '& .MuiTabs-flexContainer': { height: '100%' } }}
-					onChange={(_, value) => navigate({ to: value as MapTabRoute })}
-					orientation="vertical"
-				>
-					{/* PostAddIcon is a placeholder icon */}
-					<MapTabStyled
-						data-testid="tab-observation"
-						icon={<PostAddIcon />}
-						value={'/tab1'}
-					/>
-					<Box flexGrow={1} />
-
-					<MapTabStyled
-						icon={<SettingsIcon />}
-						label={
-							<Text style={{ fontSize: 10 }} kind="title">
-								{formatMessage(m.setting)}
-							</Text>
-						}
-						value={'/tab2'}
-					/>
-					<MapTabStyled
-						icon={<InfoOutlinedIcon />}
-						label={
-							<Text style={{ fontSize: 10 }} kind="title">
-								{formatMessage(m.about)}
-							</Text>
-						}
-						value={'/tab2'}
-					/>
-				</Tabs>
-
-				<Box
-					sx={{
+				<Tabs />
+				<div
+					style={{
 						width: 300,
 						borderLeftColor: VERY_LIGHT_GREY,
 						borderLeftWidth: '1px',
@@ -88,27 +32,11 @@ export function MapLayout() {
 					<Suspense fallback={<CircularProgress />}>
 						<Outlet />
 					</Suspense>
-				</Box>
+				</div>
 			</Paper>
 			<Suspense fallback={<CircularProgress />}>
 				<div>map component here</div>
 			</Suspense>
 		</Container>
 	)
-}
-
-type TabProps = React.ComponentProps<typeof Tab>
-
-type MapTabRoute = {
-	[K in keyof FileRoutesById]: K extends `${'/(MapTabs)/_Map'}${infer Rest}`
-		? Rest extends ''
-			? never
-			: `${Rest}`
-		: never
-}[keyof FileRoutesById]
-
-type MapTabProps = Omit<TabProps, 'value'> & { value: MapTabRoute }
-
-function MapTab(props: MapTabProps) {
-	return <Tab {...props} />
 }
